Move Postgres server impl to own mod
